refactor: extract mutation logic and state
